image-shell now accepts multiple commands
These will be run in order, without parallel with the first waiting to exit
before the second runs in a seperate chroot. signal passing, paralellism
and exit code status passing are not supported at this time.

"Quotes" will allow you run parameters for a command, the quoted content
will be passed to the shell as if it was typed.

Debian patching now works